+++ app/Http/Controllers/FacilityController.php
@@ -2,15 +2,11 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\StoreFacilityRequest;
+use App\Models\{Carrier, Facility, User,};
+use App\Services\AlertService;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
-use App\Models\{
-    User,
-    Facility,
-    Carrier,
-};
-use App\Services\AlertService;
-use App\Http\Requests\StoreFacilityRequest;
 
 class FacilityController extends Controller
 {
@@ -20,10 +16,10 @@ class FacilityController extends Controller
     protected $carrier;
 
     public function __construct(
-        User $user,
-        Facility $facilities,
+        User         $user,
+        Facility     $facilities,
         AlertService $alertService,
-        Carrier $carrier
+        Carrier      $carrier
     )
     {
         $this->user = $user;
@@ -35,10 +31,12 @@ class FacilityController extends Controller
     public function index()
     {
         $user = User::findOrFail(Auth::user()->id);
+        $this->authorize('isAdmin', $user);
+
         $facilities = Facility::where('deleted_at', null)
-            ->whereHas('carrier', function ($query) {
-                $query->where('deleted_at', null);
-            })->orderBy('id')->get();
+        ->whereHas('carrier', function ($query) {
+            $query->where('deleted_at', null);
+        })->orderBy('id')->get();
 
         return view('noc.facilities.index', compact('user', 'facilities'));
     }
@@ -46,6 +44,8 @@ class FacilityController extends Controller
     public function search(Request $request)
     {
         $user = User::findOrFail(Auth::user()->id);
+        $this->authorize('isAdmin', $user);
+
         $search = $request->input('search');
 
         if ($search === null || $search === '') {
@@ -65,6 +65,9 @@ class FacilityController extends Controller
 
     public function create()
     {
+        $user = Auth::user();
+        $this->authorize('isAdmin', $user);
+
         $carriers = Carrier::where('deleted_at', null)->get();
 
         return view('noc.facilities.create', compact('carriers'));
@@ -72,19 +75,21 @@ class FacilityController extends Controller
 
     public function store(StoreFacilityRequest $request)
     {
+        $user = Auth::user();
+        $this->authorize('isAdmin', $user);
+
         $request = $request->validated();
         $carrier = Carrier::where('deleted_at', null)->findOrFail($request['carrier_id']);
 
-        try{
+        try {
             $facility = $carrier->facilities()->create($request);
             $this->alertService->alert('success', 'Unidade cadastrada com sucesso!');
 
             return redirect()->route('noc.facilities.index');
-        }
-        catch(PDOException $e) {
+        } catch (PDOException $e) {
             $response['status'] = 0;
-            $response['msg']    = "Ops, algo inesperado aconteceu...";
-            $response['error']  = $e->getMessage();
+            $response['msg'] = "Ops, algo inesperado aconteceu...";
+            $response['error'] = $e->getMessage();
         }
 
         return $response;
@@ -92,6 +97,9 @@ class FacilityController extends Controller
 
     public function destroy(Facility $facility)
     {
+        $user = Auth::user();
+        $this->authorize('isAdmin', $user);
+
         $facility->delete();
         $this->alertService->alert('success', 'Unidade excluída com sucesso!');
